+++ games/stonecraft_game/mods/moreblocks/aliases.lua
@@ -19,6 +19,7 @@ minetest.register_alias("moreblocks:stone_bricks", "default:stonebrick")
 minetest.register_alias("moreblocks:stonebrick", "default:stonebrick")
 minetest.register_alias("moreblocks:junglewood", "default:junglewood")
 minetest.register_alias("moreblocks:jungle_wood", "default:junglewood")
+minetest.register_alias("moreblocks:fence_junglewood", "default:fence_junglewood")
 minetest.register_alias("moreblocks:fence_jungle_wood", "default:fence_junglewood")
 minetest.register_alias("moreblocks:jungle_stick", "default:stick")
 
@@ -31,13 +32,13 @@ minetest.register_alias("moreblocks:horizontaljungletree", "moreblocks:horizonta
 minetest.register_alias("moreblocks:stonesquare", "moreblocks:stone_tile")
 minetest.register_alias("moreblocks:circlestonebrick", "moreblocks:circle_stone_bricks")
 minetest.register_alias("moreblocks:ironstonebrick", "moreblocks:iron_stone_bricks")
-minetest.register_alias("moreblocks:fence_junglewood", "moreblocks:fence_jungle_wood")
 minetest.register_alias("moreblocks:coalstone", "moreblocks:coal_stone")
 minetest.register_alias("moreblocks:ironstone", "moreblocks:iron_stone")
 minetest.register_alias("moreblocks:woodtile", "moreblocks:wood_tile")
 minetest.register_alias("moreblocks:woodtile_full", "moreblocks:wood_tile_full")
 minetest.register_alias("moreblocks:woodtile_centered", "moreblocks:wood_tile_centered")
-minetest.register_alias("moreblocks:woodtile_up", "moreblocks:wood_tile_up")
+minetest.register_alias("moreblocks:woodtile_up", "moreblocks:wood_tile_offset")
+minetest.register_alias("moreblocks:wood_tile_up", "moreblocks:wood_tile_offset")
 minetest.register_alias("moreblocks:woodtile_down", "moreblocks:wood_tile_down")
 minetest.register_alias("moreblocks:woodtile_left", "moreblocks:wood_tile_left")
 minetest.register_alias("moreblocks:woodtile_right", "moreblocks:wood_tile_right")
@@ -56,6 +57,8 @@ minetest.register_alias("moreblocks:emptybookshelf", "moreblocks:empty_bookshelf
 minetest.register_alias("moreblocks:junglestick", "moreblocks:jungle_stick")
 minetest.register_alias("moreblocks:splitstonesquare","moreblocks:split_stone_tile")
 minetest.register_alias("moreblocks:allfacestree","moreblocks:all_faces_tree")
+minetest.register_alias("moreblocks:empty_bookshelf","moreblocks:empty_shelf")
+minetest.register_alias("moreblocks:split_stone_tile_alt","moreblocks:checker_stone_tile")
 
 -- ABM for horizontal trees (fix facedir):
 local horizontal_tree_convert_facedir = {7, 12, 9, 18}
@@ -75,5 +78,28 @@ minetest.register_abm({
 			name = node.name,
 			param2 = horizontal_tree_convert_facedir[node.param2 + 1]
 		})
+	end,
+})
+
+minetest.register_lbm({
+	name = "moreblocks:reduce_wood_tile_redundancy",
+	nodenames = {
+		"moreblocks:wood_tile_left",
+		"moreblocks:wood_tile_down",
+		"moreblocks:wood_tile_right",
+		"moreblocks:wood_tile_flipped",
+	},
+	action = function(pos, node)
+		if node.name:find("left") then
+			minetest.set_node(pos, {name = "moreblocks:wood_tile_offset", param2=1})
+		elseif node.name:find("down") then
+			minetest.set_node(pos, {name = "moreblocks:wood_tile_offset", param2=2})
+		elseif node.name:find("right") then
+			minetest.set_node(pos, {name = "moreblocks:wood_tile_offset", param2=3})
+		else -- wood_tile_flipped
+			minetest.set_node(pos, {name = "moreblocks:wood_tile", param2=1})
+		end
+		minetest.log('action', "LBM replaced " .. node.name ..
+			" at " .. minetest.pos_to_string(pos))
 	end,
 })